06-Visual Analytics with Folium
The launch success rate may depend on many factors such as payload mass, orbit type, and so on. It may also depend on the location and proximities of a launch site, i.e., the initial position of rocket trajectories. On this Stage of the project, we will answer the following points: 
Mark all launch sites on a map
Mark the success/failed launches for each site on the map
Calculate the distances between a launch site to its proximities